// src/fetch-apis/models/GetRecipeResponce.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from "../runtime";
import {
  type Image,
  ImageFromJSON,
  ImageFromJSONTyped,
  ImageToJSON,
} from "./Image";
import {
  type Ingredient,
  IngredientFromJSON,
  IngredientFromJSONTyped,
  IngredientToJSON,
} from "./Ingredient";
import {
  type Instruction,
  InstructionFromJSON,
  InstructionFromJSONTyped,
  InstructionToJSON,
} from "./Instruction";
import {
  type Seasoning,
  SeasoningFromJSON,
  SeasoningFromJSONTyped,
  SeasoningToJSON,
} from "./Seasoning";

/**
 *
 * @export
 * @interface GetRecipeResponce
 */
export interface GetRecipeResponce {
  /**
   *
   * @type {number}
   * @memberof GetRecipeResponce
   */
  recipeId: number;
  /**
   *
   * @type {Date}
   * @memberof GetRecipeResponce
   */
  createdAt: Date;
  /**
   *
   * @type {Date}
   * @memberof GetRecipeResponce
   */
  updatedAt: Date;
  /**
   *
   * @type {string}
   * @memberof GetRecipeResponce
   */
  name: string;
  /**
   *
   * @type {string}
   * @memberof GetRecipeResponce
   */
  comment: string;
  /**
   *
   * @type {Ingredient}
   * @memberof GetRecipeResponce
   */
  ingredient: Ingredient;
  /**
   *
   * @type {Seasoning}
   * @memberof GetRecipeResponce
   */
  seasoning: Seasoning;
  /**
   *
   * @type {Instruction}
   * @memberof GetRecipeResponce
   */
  instruction: Instruction;
  /**
   *
   * @type {Image}
   * @memberof GetRecipeResponce
   */
  image: Image;
}

export function GetRecipeResponceFromJSON(json: any): GetRecipeResponce {
  return GetRecipeResponceFromJSONTyped(json, false);
}

export function GetRecipeResponceFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): GetRecipeResponce {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    recipeId: json["recipe_id"],
    createdAt: new Date(json["created_at"]),
    updatedAt: new Date(json["updated_at"]),
    name: json["name"],
    comment: json["comment"],
    ingredient: IngredientFromJSON(json["ingredient"]),
    seasoning: SeasoningFromJSON(json["seasoning"]),
    instruction: InstructionFromJSON(json["instruction"]),
    image: ImageFromJSON(json["image"]),
  };
}

export function GetRecipeResponceToJSON(value?: GetRecipeResponce | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    recipe_id: value.recipeId,
    created_at: value.createdAt.toISOString(),
    updated_at: value.updatedAt.toISOString(),
    name: value.name,
    comment: value.comment,
    ingredient: IngredientToJSON(value.ingredient),
    seasoning: SeasoningToJSON(value.seasoning),
    instruction: InstructionToJSON(value.instruction),
    image: ImageToJSON(value.image),
  };
}